requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
google-generativeai
faiss-cpu
python-dotenv
